### 2025-11-06 #9373 @editorialbot
[PRE REVIEW]: climepi: A Python package for modeling climate suitability for vector-borne diseases
**Submitting author:** @will-s-hart (<a href="http://orcid.org/0000-0002-2504-6860">William Hart</a>)
**Repository:** https://github.com/idm-oxford/climate-epidemics
**Branch with paper.md** (empty if default branch):
**Version:** v0.6.5
