### 2025-12-22 2cda024
Update dependencies/imports for new arviz API
