### 2025-12-22 36a2859
Update dependencies/imports for new arviz API
