### 2025-12-22 a32b18b
Resolve small bugs/type checking issues from dependency updates
