+++ climepi/_core.py
@@ -533,10 +533,12 @@ class ClimEpiDatasetAccessor:
             ds_var_decomp = ds_var_decomp / ds_var_decomp.sum(dim="source")
         # Copy and update attributes and bounds
         ds_var_decomp = add_bnds_from_other(ds_var_decomp, self._obj)
-        ds_var_decomp.attrs = self._obj.attrs
+        ds_var_decomp.attrs = self._obj.attrs.copy()
         if fraction:
             for data_var_curr in data_var_list:
                 ds_var_decomp[data_var_curr].attrs["long_name"] = "Fraction of variance"
+                if "units" in ds_var_decomp[data_var_curr].attrs:
+                    del ds_var_decomp[data_var_curr].attrs["units"]
         else:
             ds_var_decomp = add_var_attrs_from_other(
                 ds_var_decomp, self._obj, var=data_var_list
