### 2025-12-23 8f20b7e
Update climepi/_core.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
