### 2025-12-23 fb71f8c
Refactor data variable type hints to use Hashable (following xarray)
