+++ climepi/_core.py
@@ -4,7 +4,7 @@ Core module for the climepi package.
 Contains the ClimEpiDatasetAccessor class for xarray datasets.
 """
 
-from typing import Any, Literal, overload
+from typing import Any, Hashable, Literal, overload
 
 import geopy
 import geoviews.feature as gf
@@ -178,7 +178,7 @@ class ClimEpiDatasetAccessor:
 
     def temporal_group_average(
         self,
-        data_var: str | list[str] | None = None,
+        data_var: Hashable | list[Hashable] | None = None,
         frequency: Literal["daily", "monthly", "yearly"] = "yearly",
         **kwargs: Any,
     ) -> xr.Dataset:
@@ -189,7 +189,7 @@ class ClimEpiDatasetAccessor:
 
         Parameters
         ----------
-        data_var : str or list, optional
+        data_var : Hashable or list of Hashable, optional
             Name(s) of the data variable(s) to compute the group average for. If not
             provided, all non-bounds data variables will be used.
         frequency : str, optional
@@ -246,7 +246,7 @@ class ClimEpiDatasetAccessor:
         return ds_m
 
     def yearly_average(
-        self, data_var: str | list[str] | None = None, **kwargs: Any
+        self, data_var: Hashable | list[Hashable] | None = None, **kwargs: Any
     ) -> xr.Dataset:
         """
         Compute the yearly mean of a data variable.
@@ -255,7 +255,7 @@ class ClimEpiDatasetAccessor:
 
         Parameters
         ----------
-        data_var : str or list, optional
+        data_var : Hashable or list of Hashable, optional
             Name(s) of the data variable(s) to compute the yearly mean for. If not
             provided, all non-bounds data variables will be used.
         **kwargs : dict, optional
@@ -272,7 +272,7 @@ class ClimEpiDatasetAccessor:
         )
 
     def monthly_average(
-        self, data_var: str | list[str] | None = None, **kwargs: Any
+        self, data_var: Hashable | list[Hashable] | None = None, **kwargs: Any
     ) -> xr.Dataset:
         """
         Compute the monthly mean of a data variable.
@@ -281,7 +281,7 @@ class ClimEpiDatasetAccessor:
 
         Parameters
         ----------
-        data_var : str or list, optional
+        data_var : Hashable or list of Hashable, optional
             Name(s) of the data variable(s) to compute the monthly mean for. If not
             provided, all non-bounds data variables will be used.
         **kwargs : dict, optional
@@ -300,7 +300,7 @@ class ClimEpiDatasetAccessor:
 
     def yearly_portion_suitable(
         self,
-        suitability_var_name: str | None = None,
+        suitability_var_name: Hashable | None = None,
         suitability_threshold: float = 0,
     ) -> xr.Dataset:
         """
@@ -312,7 +312,7 @@ class ClimEpiDatasetAccessor:
 
         Parameters
         ----------
-        suitability_var_name : str, optional
+        suitability_var_name : Hashable, optional
             Name of the suitability variable to use. If not provided, the method will
             attempt to automatically select a suitable variable.
         suitability_threshold : float, optional
@@ -370,7 +370,7 @@ class ClimEpiDatasetAccessor:
 
     def ensemble_stats(
         self,
-        data_var: str | list[str] | None = None,
+        data_var: Hashable | list[Hashable] | None = None,
         uncertainty_level: float = 90,
         internal_variability_method: Literal["direct", "polyfit", "splinefit"]
         | None = None,
@@ -382,7 +382,7 @@ class ClimEpiDatasetAccessor:
 
         Parameters
         ----------
-        data_var : str or list, optional
+        data_var : Hashable or list of Hashable, optional
[…]
