### 2025-12-23 c22e241
Simplify list comprehension for non-bound data variables in utils.py
