### 2025-12-23 031c0f5
Use catalogs from active git branch when running tests
