+++ pyproject.toml
@@ -64,6 +64,7 @@ pandas-stubs = "*"
 types-requests = "*"
 pytest = "*"
 pytest-cov = "*"
+gitpython = "*"
 jupyter_sphinx = "*"
 nbsphinx = "*"
 pydata-sphinx-theme = "*"
