### 2025-12-23 4948554
Use fetch-depth 0 in repo checkout
