+++ .github/workflows/run_tests.yml
@@ -17,6 +17,8 @@ jobs:
     steps:
     - name: Checkout repository
       uses: actions/checkout@v4
+      with:
+        fetch-depth: 0
     - name: Setup pixi
       uses: prefix-dev/setup-pixi@v0.8.14
       with:
