### 2025-12-23 16939a0
Fix patch decorator to use side_effect for _get_data_version in ARISEDataGetter tests
