### 2025-12-23 fd196d3
Don't try to infer git branch
