### 2025-12-29 b9bf0eb
Refactor arviz import to use the standard import statement No need to use conditional import as az.extract() API not due to change
