### 2025-12-29 9030e7a
Use arviz-base/stats/plots namespaces Package only needs arviz-base (arviz-stats and arviz-plots used in example notebooks so included in dev environment) Change also made to ensure fit_temperature_response() returns an xarray DataTree regardless of pymc version (currently, arviz InferenceData returned by pymc.sample, but this is likely to change)
