+++ pyproject.toml
@@ -24,10 +24,7 @@ channels = ["conda-forge", "microsoft"]
 platforms = ["linux-64", "osx-arm64", "osx-64", "win-64"]
 
 [tool.pixi.dependencies]
-arviz = ">=0.19" # first version allowing v1-like API via arviz.preview
 arviz-base = "*"
-arviz-plots="*"
-arviz-stats = "*"
 bottleneck = "*"
 dask = "*"
 datashader = "*"                 #dependency of hvplot but not always installed automatically
@@ -60,6 +57,8 @@ xcdat = "*"
 climepi = { path = ".", editable = true }
 
 [tool.pixi.feature.dev.dependencies]
+arviz-plots="*" # used in example notebooks
+arviz-stats = "*" # used in example notebooks
 ruff = "*"
 mypy = "*"
 scipy-stubs = "*"
