### 2025-12-29 a808c2e
Add test for non-string data variable in _ensemble_mean_var_polyfit
