### 2025-12-29 84d75df
Update pixi setup action to version 0.9.3
