+++ .github/workflows/run_tests.yml
@@ -18,7 +18,7 @@ jobs:
     - name: Checkout repository
       uses: actions/checkout@v4
     - name: Setup pixi
-      uses: prefix-dev/setup-pixi@v0.8.14
+      uses: prefix-dev/setup-pixi@v0.9.3
       with:
         cache: true
     - name: Lint
