### 2025-12-29 1faa5b0
Add pixi environment and test action using python 3.11 (min required version)
- Introduced a new feature group for Python 3.11 dependencies.
- Specified Python version constraint for the new feature group.
- Added a new task for running tests under the Python 3.11 feature group.
- Updated the environments section to include the new Python 3.11 feature group.
