+++ pyproject.toml
@@ -79,8 +79,12 @@ versioneer = "==0.29"
 [tool.pixi.feature.dev.pypi-dependencies]
 pytest-playwright = "*" # conda version giving errors in Windows
 
+[tool.pixi.feature.py311.dependencies]
+python = "3.11.*"
+
 [tool.pixi.environments]
 default = { features = ["dev"], solve-group = "default" }
+py311 = { features = ["py311"], solve-group = "py311" }
 
 [tool.pixi.feature.dev.tasks]
 app = "python -m climepi.app"
@@ -91,6 +95,9 @@ test = "pytest"
 docs = "python -m sphinx -T -b html -d docs/_build/doctrees -D language=en docs docs/_build/html"
 docs-rtd = "python -m sphinx -T -b html -d docs/_build/doctrees -D language=en docs $READTHEDOCS_OUTPUT/html"
 remake-examples = "python -c 'from climepi.climdata._examples import _make_all_examples; _make_all_examples(force_remake=True)'"
+
+[tool.pixi.feature.py311.tasks]
+test = "pytest"
 
 [tool.setuptools.packages.find]
 include = ["climepi*"]
