### 2025-12-29 4624069
Update actions/checkout to version v6 in CI workflow
