+++ .github/workflows/run_tests.yml
@@ -17,7 +17,7 @@ jobs:
     runs-on: ${{ matrix.os }}
     steps:
     - name: Checkout repository
-      uses: actions/checkout@v4
+      uses: actions/checkout@v6
     - name: Setup pixi
       uses: prefix-dev/setup-pixi@v0.9.3
       with:
