### 2025-12-29 4b5df42
Use include pattern to more cleanly specify os/python combinations
