### 2025-12-29 e650dd8
Don't make tasks feature-dependent
