### 2025-12-29 e16ff92
Refactor CI workflow to specify environment for linting, type checking, and testing steps
