+++ .github/workflows/run_tests.yml
@@ -31,13 +31,13 @@ jobs:
         cache: true
         environments: ${{ matrix.environment }}
     - name: Lint
-      run: pixi run lint-gh-actions
+      run: pixi run -e "${{ matrix.environment }}" lint-gh-actions
     - name: Type check
-      run: pixi run type-check
+      run: pixi run -e "${{ matrix.environment }}" type-check
     - name: Install browsers for Playwright tests
-      run: pixi run playwright install chromium --with-deps
+      run: pixi run -e "${{ matrix.environment }}" playwright install chromium --with-deps
     - name: Test with pytest
-      run: pixi run test
+      run: pixi run -e "${{ matrix.environment }}" test
     - name: Upload coverage to Codecov
       if: ${{ !cancelled() }}
       uses: codecov/codecov-action@v5
@@ -49,5 +49,5 @@ jobs:
       uses: codecov/codecov-action@v5
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
-        flags: ${{ matrix.os }}, ${{ matrix.environment }}
+        flags: ${{ matrix.os }},${{ matrix.environment }}
         report_type: test_results
