### 2025-12-29 18af2fe
Only lint and type check for ubuntu with default environment
