+++ .github/workflows/run_tests.yml
@@ -31,8 +31,10 @@ jobs:
         cache: true
         environments: ${{ matrix.environment }}
     - name: Lint
+      if: ${{ matrix.environment == 'default' && matrix.os == 'ubuntu-latest' }}
       run: pixi run -e "${{ matrix.environment }}" lint-gh-actions
     - name: Type check
+      if: ${{ matrix.environment == 'default' && matrix.os == 'ubuntu-latest' }}
       run: pixi run -e "${{ matrix.environment }}" type-check
     - name: Install browsers for Playwright tests
       run: pixi run -e "${{ matrix.environment }}" playwright install chromium --with-deps
