### 2025-12-29 39ac848
Update py311 environment to include 'dev' dependencies (needed for CI)
