### 2025-12-29 19f4949
Use TypedDict from typing-extensions for python 3.11 compatibility
