### 2025-12-29 b623ff3
Add dependabot configuration for GitHub Actions updates
