+++ .github/dependabot.yml
@@ -1,0 +1,10 @@
+version: 2
+updates:
+  - package-ecosystem: github-actions
+    directory: /
+    schedule:
+      interval: monthly #
+    groups:
+      dependencies:
+        patterns:
+          - '*'
