+++ .github/dependabot.yml
@@ -3,7 +3,7 @@ updates:
   - package-ecosystem: github-actions
     directory: /
     schedule:
-      interval: monthly #
+      interval: monthly
     groups:
       dependencies:
         patterns:
