### 2025-12-29 0663a2b
Replace setuptools and versioneer with hatchling and version-pioneer
