### 2025-12-29 7b02a78
Add version_pioneer to mypy overrides
