### 2025-12-29 b3d9ff1
Refactor version retrieval in _get_data_version function
