### 2025-12-29 3fbe23a
Update pyproject.toml
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
