### 2025-12-29 da3588d
Remove unneeded package data includes
