+++ pyproject.toml
@@ -136,7 +136,6 @@ convention = "numpy"
 
 [tool.mypy]
 check_untyped_defs = true
-# ignore_missing_imports = true
 warn_unused_ignores = true
 warn_redundant_casts = true
 warn_unused_configs = true
