### 2026-01-02 c89b771
Add scheduled workflow to run tests every Monday at 3 AM
