+++ .github/workflows/run_tests.yml
@@ -6,6 +6,8 @@ on:
   pull_request:
     branches: [ "main" ]
   workflow_dispatch:
+  schedule:
+    - cron: "0 3 * * 1"
 
 jobs:
   build:
