### 2026-01-02 0724c8f
Add test runs for each of python 3.11, 3.12, 3.13 and 3.14
