+++ .github/workflows/run_tests.yml
@@ -19,6 +19,12 @@ jobs:
             environment: default
           - os: ubuntu-latest
             environment: py311
+          - os: ubuntu-latest
+            environment: py312
+          - os: ubuntu-latest
+            environment: py313
+          - os: ubuntu-latest
+            environment: py314
           - os: windows-latest
             environment: default
           - os: macos-latest
