### 2026-01-02 ca5a366
Increase max parallel jobs from 5 to 7 in CI workflow
