### 2026-01-02 32d4537
Change config following pixi example from Read the Docs docs https://docs.readthedocs.com/platform/stable/build-customization.html
