+++ .readthedocs.yaml
@@ -1,10 +1,16 @@
 version: 2
 
 build:
-  os: "ubuntu-22.04"
-  tools:
-    python: mambaforge-latest
-  commands:
-    - mamba install -c conda-forge -c nodefaults pixi
-    - pixi install
-    - pixi run docs-rtd
+   os: ubuntu-24.04
+   tools:
+       python: "latest"
+   jobs:
+      create_environment:
+         - asdf plugin add pixi
+         - asdf install pixi latest
+         - asdf global pixi latest
+      install:
+         - pixi install --locked
+      build:
+         html:
+            - pixi run docs-rtd
