### 2026-01-02 a2ac6b7
Pin sphinx<9 due to sphinx-autosummary-accessors incompatibility https://github.com/xarray-contrib/sphinx-autosummary-accessors/issues/165
