### 2026-01-02 bb92cca
Replace deprecated jupyter_sphinx alias with jupyter-sphinx
