+++ pyproject.toml
@@ -68,7 +68,7 @@ types-requests = "*"
 pytest = "*"
 pytest-cov = "*"
 sphinx = "<9.0.0"  # https://github.com/xarray-contrib/sphinx-autosummary-accessors/issues/165
-jupyter_sphinx = "*"
+jupyter-sphinx = "*"
 nbsphinx = "*"
 pydata-sphinx-theme = "*"
 sphinx-autosummary-accessors = "*"
