### 2026-01-02 93723e4
Remove coverage/test-results flags from Codecov uploads in CI workflow
