### 2026-01-07 1c649a0
Use vendored _version.py
