### 2026-01-07 2358ef1
Prefer directory-name based version when available, handle cases where periods changed to underscores Fixes version inference during conda build (previously, git tags were taken from a parent repo containing the downloaded code)
