### 2026-01-07 1117dd2
Update climepi/_version.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
