### 2026-01-07 fd808d9
Update climepi/_version.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
