### 2026-01-07 7cf506d
Fix type checking issues
