### 2026-01-07 c40ddc2
Revert to non-vendored version script
