+++ climepi/_version.py
@@ -1,1111 +1,20 @@
-#!/usr/bin/env python3
-"""
-Generate a version number from Git tags (e.g. tag "v1.2.3" and 4 commits -> "1.2.3+4.g123abcdef" in "pep440" style).
-
-This "versionscript" may be replaced with a much shorter "versionfile" in distribution tarballs
-(built by `uv build`, `pyproject-build`) that just contains one method:
-`def get_version_dict() -> VersionDict: return {"version": "0.1.0", ...}`.
-
-Refactored from Versioneer's _version.py.
-
-Note:
-    - Should be compatible with python 3.8+ without any dependencies.
-        - (For dev) Avoid importing third-party libraries, including version-pioneer itself
-          because this file can get vendored into other projects.
-        - (For user) Once the script is vendored, and you want to customise it, of course you can
-          import other libraries and add those in build-time dependencies.
-    - This file is usually located at `src/my_package/_version.py`.
-    - `src/my_package/__init__.py` should define `__version__ = get_version_dict()["version"]` by importing this module.
-    - It should also be able to be run as a script to print the version info in json format.
-    - It is often `exec`-uted and `get_version_dict()` is evaluated from this file.
-        - Using `from __future__ import ...` with dataclasses makes it hard to "exec" this file,
-          so you MUST NOT use both here.
-            - See https://github.com/mkdocs/mkdocs/issues/3141
-              https://github.com/pypa/hatch/issues/1863
-              https://github.com/sqlalchemy/alembic/issues/1419
-            - You need to put the module in `sys.modules` before executing
-              because dataclass will look for the type there.
-            - While this can be fixed, it's a common gotcha and I expect that
-              some build backends or tools will be buggy.
-            - For now, we don't future import. In dataclass definition,
-              we use `typing.Optional` instead of `| None`
-              until we drop support for Python 3.9.
-"""
-
-# ruff: noqa: T201 FA100
-
-import contextlib
-import errno
-import functools
-import os
-import re
-import subprocess
-import sys
-from collections.abc import Iterable
-from dataclasses import dataclass
-from email.parser import Parser
-from enum import Enum
-from os import PathLike
+# Example _version.py, completely non-vendored.
 from pathlib import Path
-from typing import Any, Literal, Optional, TypedDict, TypeVar, Union
 
+from version_pioneer.api import get_version_dict_wo_exec
 
-class VersionStyle(str, Enum):
-    pep440 = "pep440"
-    pep440_master = "pep440-master"
-    pep440_branch = "pep440-branch"
-    pep440_pre = "pep440-pre"
-    pep440_post = "pep440-post"
-    pep440_post_branch = "pep440-post-branch"
-    git_describe = "git-describe"
-    git_describe_long = "git-describe-long"
-    digits = "digits"
 
-
-VERSION_STYLE_TYPE = TypeVar(
-    "VERSION_STYLE_TYPE",
-    Literal[
-        "pep440",
-        "pep440-master",
-        "pep440-branch",
-        "pep440-pre",
-        "pep440-post",
-        "pep440-post-branch",
-        "git-describe",
-        "git-describe-long",
-        "digits",
-    ],
-    VersionStyle,
-)
-
-
-# ┌──────────────────────────────────────────┐
-# │     Modify the configuration below.      │
-# └──────────────────────────────────────────┘
-@dataclass(frozen=True)
-class VersionPioneerConfig:
-    style: VersionStyle = VersionStyle.pep440
-    tag_prefix: str = "v"
-    # if there is no .git, like it's a source tarball downloaded from GitHub Releases,
-    # find version from the name of the parent directory.
[…]
