### 2026-01-07 7bb4e33
Try to introduce git archive version substitution
