### 2026-01-08 1ec60bf
Add license information to pyproject.toml
