### 2026-01-08 8a8c83b
Update license format in pyproject.toml
