+++ pyproject.toml
@@ -9,7 +9,7 @@ authors = [{ name = "William Hart", email = "william.hart@maths.ox.ac.uk" }]
 description = "Combining climate data and epidemiological models"
 readme = "README.md"
 requires-python = ">=3.11"
-license = { text = "GPL-3.0-only" }
+license = "GPL-3.0-only"
 classifiers = [
   "Programming Language :: Python :: 3",
   "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
