### 2026-01-15 ad01b48
Reformat paper following new JOSS specification
