### 2026-01-15 2d0a5b0
Small wording tweak
