+++ paper/paper.md
@@ -16,7 +16,7 @@ authors:
 affiliations:
  - name: Mathematical Institute, University of Oxford, Oxford, OX2 6GG, UK
    index: 1
-date: 15 January 2025
+date: 15 January 2026
 bibliography: paper.bib
 ---
 
