### 2026-01-15 c1e454b
Add details about use of xarray to design section
