### 2026-01-16 a860e05
Style changes to AI disclosure
