### 2026-02-11 ee495cb
Move figure up to reduce white space
