+++ paper/paper.md
@@ -62,7 +62,21 @@ https://docs.xarray.dev/en/stable/user-guide/dask.html) with [Dask](https://dask
 enabling lazy evaluation and parallel computation on larger-than-memory datasets.
 climepi is designed with a modular structure, comprising three main components
 corresponding to different stages of a typical climate-VBD suitability modeling
-workflow:
+workflow (\autoref{fig:schematic}):
+
+![Schematic illustrating typical workflows using climepi. (1) The `climdata` subpackage
+can be used to access climate data from an external source, or to load a pre-defined
+example climate dataset. (2) The `epimod` subpackage can be used to define a mechanistic
+model describing climate suitability for VBD, infer temperature response curves of trait 
+parameters using laboratory data (left plot), and determine the overall temperature
+dependence of the suitability metric (right plot). Pre-defined epidemiological models
+from the literature are also available as built-in examples. (3) The `climepi` accessor
+for xarray datasets can be used to combine a climate dataset with an epidemiological
+model of climate suitability for VBD to obtain suitability projections, to assess the
+contributions of different sources of climate uncertainty, and to visualize results. See
+the package [documentation](
+https://climate-epidemics.readthedocs.io/en/stable/gallery.html) for detailed usage
+examples.\label{fig:schematic}](schematic.pdf)
 
 1. Climate data (`climdata`) subpackage: enables users to access climate projection data
    from different sources through a single interface. Rather than providing
@@ -79,21 +93,6 @@ workflow:
    https://www.cesm.ucar.edu/community-projects/lens2) (LENS2) [@rodgers_ubiquity_2021],
    which provides 100 ensemble members for analyzing internal (natural) climate
    variability.
-
-![Schematic illustrating typical workflows using climepi. (1) The `climdata` subpackage
-can be used to access climate data from an external source, or to load a pre-defined
-example climate dataset. (2) The `epimod` subpackage can be used to define a mechanistic
-model describing climate suitability for VBD, infer temperature response curves of trait 
-parameters using laboratory data (left plot), and determine the overall temperature
-dependence of the suitability metric (right plot). Pre-defined epidemiological models
-from the literature are also available as built-in examples. (3) The `climepi` accessor
-for xarray datasets can be used to combine a climate dataset with an epidemiological
-model of climate suitability for VBD to obtain suitability projections, to assess the
-contributions of different sources of climate uncertainty, and to visualize results. See
-the package [documentation](
-https://climate-epidemics.readthedocs.io/en/stable/gallery.html) for detailed usage
-examples.](schematic.pdf)
-
 2. Epidemiological model (`epimod`) subpackage: provides classes and methods for models
    of climate suitability for VBD, in which a suitability metric (e.g., the basic
    reproduction number, $R_0$) is defined as a function of temperature and/or
