### 2026-02-11 9d012bb
Add final sentence to state of the field section
