### 2026-02-12 #9373 @will-s-hart
@arfon @kthyng Apologies for the delay; I have now added the State of the Field section as requested.
