### 2026-02-12 992affe
Compute dataset before generating plot in _Plotter class Improves speed and reduces risk of param errors
