+++ climepi/app/_app_classes_methods.py
@@ -163,7 +163,7 @@ class _Plotter:
     def generate_plot(self) -> None:
         """Generate the plot."""
         self._get_ds_plot()
-        ds_plot = self._ds_plot
+        ds_plot = self._ds_plot.compute()
         assert ds_plot is not None
         plot_settings = self._plot_settings
         plot_type = plot_settings["plot_type"]
