### 2026-02-12 18240f4
Add assertion to ensure _ds_plot is not None before computing the plot Ensures type checking success
