### 2026-02-20 #9373 @chrisvoncsefalvay
> :wave: @robmoss @fabianaganem @chrisvoncsefalvay @Chilipp, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. We ask for reviews to start in the next 2 weeks and aim to finish within 4-6 weeks.

Happy to do so.
