### 2026-02-23 #10099 @will-s-hart
@willgearty thanks for this.

I noticed the version number update from the pre-review thread didn't carry through to here, so please can this be updated to the latest version, v0.6.6?
