### 2026-03-10 #10099 @chrisvoncsefalvay
I am happy to endorse this for publication with two remarks.

1) Collaborative effort: the package has pretty much [one single maintainer](https://github.com/idm-oxford/climate-epidemics/graphs/contributors). It would conflict with my obligations as a reviewer if I did not point out that this does offend the specified criterion that projects have "collaborative effort". At the same time, I would very respectfully request that this not be held against the package in any way. Climate-specific vector-borne epidemiology is not a massive field. The number of individuals qualified to constitute the 'communtiy' from which collaborators would be drawn is enormously small to begin with, and that gives us a kind of chicken/egg dilemma: publication may be impeded by lack of community participation, yet it may be the best way to raise the level of awareness that is required to garner the attention needed for such participation to eventuate. It would not be appropriate for me to discuss generalities or question the merit of this criterion (I think I see why it exists), but merely to state its impact in this present case: @will-s-hart has produced a package on a neglected field, all alone, that meets the quality criteria I would expect from a competently written open source statistical software package, be it written by a team of one or a team of a hundred. He did so in a way that did not necessitate external bug-fixing. Sometimes, the absence of issues doesn't mean lack of community participation but actually well-written software. I cannot see why any of these should now count against him. Therefore, while I am leaving the relevant box unchecked for factuality's sake, I would ask the editors to please consider my endorsement complete even without. @willgearty
2) In the interests of full disclosure: I am an alumnus (twice over at least, if not more) of the University of Oxford. By virtue of Statute III of the University, that makes me a member of the Convocation, which strictly speaking means I have an institutional affiliation I share with the author. In practice, however, this is not in any way a connection that could realistically raise even the remotest concern of a CoI. Nevertheless, I feel it is appropriate that this be so disclosed and recorded. (For the record, I don't know the author, have had no involvement w/ the Mathematical Institute and haven't exercised any rights in respect of Convocation except seconding someone for Professor of Poetry a few years ago.)
