### 2026-04-27 170aa5d
Add scheduled workflow for monthly pixi lockfile updates
