+++ .github/workflows/update_lockfile.yml
@@ -1,0 +1,33 @@
+name: Update pixi lockfile
+
+on:
+  schedule:
+    - cron: "0 3 1 * *"   # 1st of each month, 03:00 UTC
+  workflow_dispatch:
+
+jobs:
+  update-lockfile:
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write
+      pull-requests: write
+    steps:
+      - uses: actions/checkout@v6
+
+      - uses: prefix-dev/setup-pixi@v0.9.4
+        with:
+          cache: true
+
+      - name: Update lockfile
+        run: pixi update
+
+      - name: Open PR if lockfile changed
+        uses: peter-evans/create-pull-request@v7
+        with:
+          branch: update/pixi-lockfile
+          commit-message: "chore: update pixi lockfile"
+          title: "chore: monthly pixi lockfile update"
+          body: |
+            Automated monthly lockfile update. CI runs on this PR to verify compatibility.
+          labels: dependencies
+          delete-branch: true
