### 2026-04-27 8008ad8
Update pixi setup action to version 0.9.5 in lockfile update workflow
