+++ .github/workflows/update_lockfile.yml
@@ -13,14 +13,11 @@ jobs:
       pull-requests: write
     steps:
       - uses: actions/checkout@v6
-
-      - uses: prefix-dev/setup-pixi@v0.9.4
+      - uses: prefix-dev/setup-pixi@v0.9.5
         with:
           cache: true
-
       - name: Update lockfile
         run: pixi update
-
       - name: Open PR if lockfile changed
         uses: peter-evans/create-pull-request@v7
         with:
