### 2026-04-27 d6ff888
Add concurrency settings to lockfile update workflow
