### 2026-04-27 d2a9f72
Update DOI badge in README.md Old version often too slow to render
