### 2026-04-27 72e0e33
Update create-pull-request action to version 8 in lockfile update workflow
