+++ .github/workflows/update_lockfile.yml
@@ -21,7 +21,7 @@ jobs:
       - name: Update lockfile
         run: pixi update
       - name: Open PR if lockfile changed
-        uses: peter-evans/create-pull-request@v7
+        uses: peter-evans/create-pull-request@v8
         with:
           branch: update/pixi-lockfile
           commit-message: "chore: update pixi lockfile"
