### 2026-04-27 b8f9fd7
Update pixi setup action to prevent installation during lockfile update
