+++ .github/workflows/update_lockfile.yml
@@ -17,7 +17,7 @@ jobs:
       - uses: actions/checkout@v6
       - uses: prefix-dev/setup-pixi@v0.9.5
         with:
-          cache: true
+          run-install: false
       - name: Update lockfile
         run: pixi update
       - name: Open PR if lockfile changed
