### 2026-04-27 a501d8c
Add token generation step for lockfile update workflow
