+++ .github/workflows/update_lockfile.yml
@@ -20,9 +20,16 @@ jobs:
           run-install: false
       - name: Update lockfile
         run: pixi update
+      - name: Generate app token
+        id: app-token
+        uses: actions/create-github-app-token@v1
+        with:
+          app-id: ${{ secrets.APP_ID }}
+          private-key: ${{ secrets.APP_PRIVATE_KEY }}
       - name: Open PR if lockfile changed
         uses: peter-evans/create-pull-request@v8
         with:
+          token: ${{ steps.app-token.outputs.token }}
           branch: update/pixi-lockfile
           commit-message: "chore: update pixi lockfile"
           title: "chore: monthly pixi lockfile update"
