### 2026-04-27 27936ee
Update create-github-app-token action to version 3 in lockfile update workflow
