+++ .github/workflows/update_lockfile.yml
@@ -22,7 +22,7 @@ jobs:
         run: pixi update
       - name: Generate app token
         id: app-token
-        uses: actions/create-github-app-token@v1
+        uses: actions/create-github-app-token@v3
         with:
           app-id: ${{ secrets.APP_ID }}
           private-key: ${{ secrets.APP_PRIVATE_KEY }}
