### 2026-04-27 20e2dfb
Reduce permissions for update-lockfile job to enhance security
