+++ .github/workflows/update_lockfile.yml
@@ -11,8 +11,7 @@ jobs:
   update-lockfile:
     runs-on: ubuntu-latest
     permissions:
-      contents: write
-      pull-requests: write
+      contents: read
     steps:
       - uses: actions/checkout@v6
       - uses: prefix-dev/setup-pixi@v0.9.5
