### 2026-04-27 6d84fd0
Fix persist-credentials setting in checkout step for lockfile update workflow
