+++ .github/workflows/update_lockfile.yml
@@ -14,6 +14,8 @@ jobs:
       contents: read
     steps:
       - uses: actions/checkout@v6
+        with:
+          persist-credentials: false
       - uses: prefix-dev/setup-pixi@v0.9.5
         with:
           run-install: false
