### 2026-04-27 307a39e
Fix parameter name for app token generation in lockfile update workflow
