### 2026-04-27 bb89627
Fix type checking issues
