### 2026-04-27 d4b25a4
Revert changes that broke attr handling, use alt type checkiong fix
