### 2026-05-01 bb3de66
Harden threaded app test teardown
Agent-Logs-Url: https://github.com/idm-oxford/climate-epidemics/sessions/74168bc1-7610-40a6-b052-ae88a9c69420

Co-authored-by: will-s-hart <47417085+will-s-hart@users.noreply.github.com>
