### 2026-05-04 df090aa
Address PR review: safe shutdown timeout + autouse signal handler fixture
Agent-Logs-Url: https://github.com/idm-oxford/climate-epidemics/sessions/2e135a8c-a761-4e6c-8f91-d6de2170c7f1

Co-authored-by: will-s-hart <47417085+will-s-hart@users.noreply.github.com>
