### 2026-05-12 1b35815
Refactor app testing: add context manager for threaded app execution
