### 2026-05-29 e43aa96
Add quickstart to README with small example dataset
