### 2026-05-29 5606d7f
Update quickstart Include instructions for showing figures when running in a script rather than notebook
