### 2026-05-29 faa3c3e
Incorporate lazy imports to reduce import overhead Smaller improvements added via PEP 810 syntax, so will only take effect on Python 3.15 release
