### 2026-05-29 79a7b88
Fix type checking issues Tighten semantics of geocode to prevent list return, and check for geocode returning none
