+++ climepi/_geocoding.py
@@ -27,23 +27,32 @@ def _initialize_geocode() -> None:
 
 
 @lru_cache(maxsize=1000, typed=True)
-def geocode(*args: Any, **kwargs: Any) -> Location | None | list[Location | None]:
+def geocode(query: str, **kwargs: Any) -> Location | None:
     """
     Geocode an address using the Nominatim geocoder.
 
-    Uses OpenStreetMap data (https://openstreetmap.org/copyright).
+    Uses OpenStreetMap data (https://openstreetmap.org/copyright). Always returns
+    a single :class:`geopy.Location` (or ``None`` if the query cannot be resolved);
+    passing ``exactly_one=False`` is rejected.
 
     Parameters
     ----------
-    *args, **kwargs:
-        Arguments and keyword arguments passed to the Nominatim.geocode method (see
+    query : str
+        Query string passed to the Nominatim.geocode method (see
         https://geopy.readthedocs.io/en/stable/#geopy.geocoders.Nominatim.geocode).
+    **kwargs:
+        Additional keyword arguments passed to the Nominatim.geocode method.
 
     Returns
     -------
-    geopy.Location or None or list:
+    geopy.Location or None:
         Return value of the Nominatim.geocode method (see the link above).
     """
+    if kwargs.get("exactly_one", True) is False:
+        raise ValueError(
+            "geocode() always returns a single Location (or None); "
+            "'exactly_one=False' is not supported."
+        )
     _initialize_geocode()
     assert _geocode is not None, "Geocode service is not initialized."
-    return _geocode(*args, **kwargs)
+    return _geocode(query, **kwargs)
