### 2026-05-29 d78fda7
Update AI disclosure
