### 2026-05-29 8a1cee1
Add not None assertions to avoid type checking errors
