### 2026-05-29 b853704
Change panel widget parameter from name->label "name" deprecated, so replace with label and pin minimum panel version to ensure label alias available
