### 2026-05-29 4c169fa
Enforce not None return value at geocode level Avoids need for assertions/checks elsewhere
