+++ climepi/_geocoding.py
@@ -27,13 +27,13 @@ def _initialize_geocode() -> None:
 
 
 @lru_cache(maxsize=1000, typed=True)
-def geocode(query: str, **kwargs: Any) -> Location | None:
+def geocode(query: str, **kwargs: Any) -> Location:
     """
     Geocode an address using the Nominatim geocoder.
 
     Uses OpenStreetMap data (https://openstreetmap.org/copyright). Always returns
-    a single :class:`geopy.Location` (or ``None`` if the query cannot be resolved);
-    passing ``exactly_one=False`` is rejected.
+    a single :class:`geopy.Location`; passing ``exactly_one=False`` is rejected,
+    and unresolved queries raise :class:`ValueError`.
 
     Parameters
     ----------
@@ -45,14 +45,22 @@ def geocode(query: str, **kwargs: Any) -> Location | None:
 
     Returns
     -------
-    geopy.Location or None:
+    geopy.Location:
         Return value of the Nominatim.geocode method (see the link above).
+
+    Raises
+    ------
+    ValueError
+        If ``exactly_one=False`` is passed, or if the query cannot be resolved.
     """
     if kwargs.get("exactly_one", True) is False:
         raise ValueError(
-            "geocode() always returns a single Location (or None); "
+            "geocode() always returns a single Location; "
             "'exactly_one=False' is not supported."
         )
     _initialize_geocode()
     assert _geocode is not None, "Geocode service is not initialized."
-    return _geocode(query, **kwargs)
+    result = _geocode(query, **kwargs)
+    if result is None:
+        raise ValueError(f"Could not geocode query {query!r}.")
+    return result
