### 2026-05-29 144295b
Add PEP 810 lazy module declarations
