### 2026-06-02 c446b26
Bump prefix-dev/setup-pixi from 0.9.5 to 0.9.6 in the dependencies group
Bumps the dependencies group with 1 update: [prefix-dev/setup-pixi](https://github.com/prefix-dev/setup-pixi).


Updates `prefix-dev/setup-pixi` from 0.9.5 to 0.9.6
- [Release notes](https://github.com/prefix-dev/setup-pixi/releases)
- [Commits](https://github.com/prefix-dev/setup-pixi/compare/v0.9.5...v0.9.6)

---
updated-dependencies:
- dependency-name: prefix-dev/setup-pixi
  dependency-version: 0.9.6
  dependency-type: direct:production
  update-type: version-update:semver-patch
  dependency-group: dependencies
...

S
[…]
