+++ .github/workflows/run_tests.yml
@@ -32,7 +32,7 @@ jobs:
     - name: Checkout repository
       uses: actions/checkout@v6
     - name: Setup pixi
-      uses: prefix-dev/setup-pixi@v0.9.5
+      uses: prefix-dev/setup-pixi@v0.9.6
       with:
         cache: true
         environments: ${{ matrix.environment }}
